3D: Fix for 'warning C4018: '<' : signed/unsigned mismatch' in DevVGA-SVGA3d-hlp.cpp
svn:sync-xref-src-repo-rev: r140960

+++ src/VBox/Devices/Graphics/DevVGA-SVGA3d-hlp.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevVGA-SVGA3d-hlp.cpp 86589 2020-10-15 11:32:19Z knut.osmundsen@oracle.com $ */
+/* $Id: DevVGA-SVGA3d-hlp.cpp 86592 2020-10-15 14:21:21Z dmitrii.grigorev@oracle.com $ */
 /** @file
  * DevVMWare - VMWare SVGA device helpers
  */
@@ -355,7 +355,7 @@ int vmsvga3dShaderParse(SVGA3dShaderType type, uint32_t cbShaderData, uint32_t c
         Log3(("op %d, cInstLen %d\n", token.op, cInstLen));
 
         /* Must not be greater than the number of remaining tokens. */
-        ASSERT_GUEST_RETURN(cInstLen < paTokensEnd - pToken, VERR_PARSE_ERROR);
+        ASSERT_GUEST_RETURN(cInstLen < (uintptr_t)(paTokensEnd - pToken), VERR_PARSE_ERROR);
 
         /* Stop parsing if this is the SVGA3DOP_END instruction. */
         if (token.op == SVGA3DOP_END)
